requirements.txt
python-dotenv
streamlit
transformers
torch
requests
datetime
python-dotenv